fix dtype

+++ src/models/stage2_inpaint_unet_2d_condition.py
@@ -679,7 +679,7 @@ class Stage2_InapintUNet2DConditionModel(ModelMixin, ConfigMixin, UNet2DConditio
         # `Timesteps` does not contain any weights and will always return f32 tensors
         # but time_embedding might actually be running in fp16. so we need to cast here.
         # there might be better ways to encapsulate this.
-        t_emb = t_emb.to(dtype=torch.bfloat16)  # [bs, 320]
+        t_emb = t_emb.to(dtype=sample.dtype)  # [bs, 320]
 
         emb = self.time_embedding(t_emb, timestep_cond)  # [bs,1280]
 
@@ -739,7 +739,7 @@ class Stage2_InapintUNet2DConditionModel(ModelMixin, ConfigMixin, UNet2DConditio
             encoder_hidden_states = self.encoder_hid_proj(encoder_hidden_states, image_embeds)
 
         # 2. pre-process
-        sample = self.conv_in(sample.to(torch.bfloat16))
+        sample = self.conv_in(sample)
